+++ examples/demo-rollup/tests/bank/mod.rs
@@ -1,5 +1,6 @@
 mod helpers;
 mod op_rollup;
+mod operator_rollup;
 mod zk_rollup;
 
 const TOKEN_NAME: &str = "test_token";